requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
waitress==3.0.0
sqlalchemy==2.0.25
python-multipart==0.0.6
jinja2==3.1.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dateutil==2.8.2
 